+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/DialogContentConsumer.kt
@@ -1,8 +1,8 @@
 package ru.mintrocket.lib.mintpermissions.flows
 
 import androidx.activity.ComponentActivity
-import ru.mintrocket.lib.mintpermissions.flows.model.DialogRequestContent
-import ru.mintrocket.lib.mintpermissions.flows.model.DialogResult
+import ru.mintrocket.lib.mintpermissions.flows.models.DialogRequestContent
+import ru.mintrocket.lib.mintpermissions.flows.models.DialogResult
 
 interface DialogContentConsumer {
