+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/DialogsController.kt
@@ -1,7 +1,7 @@
 package ru.mintrocket.lib.mintpermissions.flows
 
-import ru.mintrocket.lib.mintpermissions.flows.model.DialogRequest
-import ru.mintrocket.lib.mintpermissions.flows.model.DialogResult
+import ru.mintrocket.lib.mintpermissions.flows.models.DialogRequest
+import ru.mintrocket.lib.mintpermissions.flows.models.DialogResult
 import ru.mintrocket.lib.mintpermissions.tools.uirequests.UiRequestController
 
 class DialogsController(
