--- flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/model/DialogRequestContent.kt
@@ -1,8 +1,0 @@
-package ru.mintrocket.lib.mintpermissions.flows.model
-
-import ru.mintrocket.lib.mintpermissions.tools.uirequests.model.UiRequest
-
-data class DialogRequestContent(
-    val request: UiRequest<DialogRequest>,
-    val content: DialogContent
-)
