--- flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/model/DialogResult.kt
@@ -1,6 +1,0 @@
-package ru.mintrocket.lib.mintpermissions.flows.model
-
-enum class DialogResult {
-    CANCEL,
-    ACTION
-}
