--- flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/model/PlainContent.kt
@@ -1,4 +1,0 @@
-package ru.mintrocket.lib.mintpermissions.flows.model
-
-class PlainContent {
-}
