+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/models/DialogRequestContent.kt
@@ -1,0 +1,8 @@
+package ru.mintrocket.lib.mintpermissions.flows.models
+
+import ru.mintrocket.lib.mintpermissions.tools.uirequests.models.UiRequest
+
+data class DialogRequestContent(
+    val request: UiRequest<DialogRequest>,
+    val content: DialogContent
+)
