+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/models/PlainContent.kt
@@ -1,0 +1,4 @@
+package ru.mintrocket.lib.mintpermissions.flows.models
+
+class PlainContent {
+}
